+++ src/main/kotlin/com/github/xeonkryptos/gitlabintegration/gitlab/util/GitlabUtil.kt
@@ -1,0 +1,13 @@
+package com.github.xeonkryptos.gitlabintegration.gitlab.util
+
+import com.intellij.openapi.diagnostic.Logger
+
+/**
+ * @author Xeonkryptos
+ * @since 06.09.2020
+ */
+object GitlabUtil {
+
+    @JvmField
+    val LOG: Logger = Logger.getInstance("gitlab")
+}